+++ src/main/java/um/tesoreria/guarani/hexagonal/guarani/alumnos/alumno/application/usecases/CreatePreuniversitarioByNroDocumentoUseCaseImpl.java
@@ -4,9 +4,9 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import um.tesoreria.guarani.hexagonal.guarani.alumnos.alumno.domain.model.AlumnoGuarani;
+import um.tesoreria.guarani.hexagonal.guarani.alumnos.alumno.domain.ports.in.CreatePersonalesByNroDocumentoUseCase;
 import um.tesoreria.guarani.hexagonal.guarani.alumnos.alumno.domain.ports.in.CreatePreuniversitarioByNroDocumentoUseCase;
-import um.tesoreria.guarani.hexagonal.guarani.alumnos.alumno.domain.ports.in.CreatePreuniversitarioUseCase;
-import um.tesoreria.guarani.hexagonal.guarani.alumnos.alumno.domain.ports.in.GetAlumnosByNroDocumentoUseCase;
+import um.tesoreria.guarani.hexagonal.guarani.alumnos.alumno.infrastructure.client.dto.CreatePersonalesResponse;
 
 import java.util.List;
 
@@ -15,23 +15,29 @@ import java.util.List;
 @RequiredArgsConstructor
 public class CreatePreuniversitarioByNroDocumentoUseCaseImpl implements CreatePreuniversitarioByNroDocumentoUseCase {
 
-    private static final int PROPUESTA_TIPO_PREUNIVERSITARIO = 204;
-
-    private final GetAlumnosByNroDocumentoUseCase getAlumnosByNroDocumentoUseCase;
-    private final CreatePreuniversitarioUseCase createPreuniversitarioUseCase;
+    private final CreatePersonalesByNroDocumentoUseCase createPersonalesByNroDocumentoUseCase;
 
     @Override
     public List<AlumnoGuarani> createPreuniversitarioByNroDocumento(String nroDocumento) {
         log.debug("\n\nCreating preuniversitario for document: {}\n\n", nroDocumento);
-        List<AlumnoGuarani> alumnos = getAlumnosByNroDocumentoUseCase.getByNroDocumento(nroDocumento);
-        
-        // Filter only pre-university students (propuestaTipo == 204)
-//        List<AlumnoGuarani> preuniversitarioAlumnos = alumnos.stream()
-//                .filter(alumno -> alumno.getPropuestaRel() != null
-//                        && Integer.valueOf(PROPUESTA_TIPO_PREUNIVERSITARIO).equals(alumno.getPropuestaRel().getPropuestaTipo()))
-//                .toList();
+        List<CreatePersonalesResponse> creados = null;
+        if ((creados = createPersonalesByNroDocumentoUseCase.createPersonalesByNroDocumento(nroDocumento)).isEmpty()) {
+            return List.of();
+        }
 
-//        return createPreuniversitarioUseCase.createPreuniversitario(preuniversitarioAlumnos);
+//        List<AlumnoGuarani> generados = new ArrayList<>();
+//        for (var persona:creados) {
+//            try {
+//                log.debug("\n\nProcessing Alumno -> {}\n\n", Jsonifier.builder(alumno.getPersonaRel()).build());
+//                AlumnoGuarani creado = alumnoGuaraniClient.createPreuniversitario(alumno);
+//                if (creado != null) {
+//                    creados.add(creado);
+//                }
+//            } catch (Exception e) {
+//                log.error("Error creating preuniversitario for alumno {}: {}", alumno.getAlumno(), e.getMessage());
+//            }
+//        }
+//        return creados;
         return null;
     }
 }