+++ src/main/java/um/tesoreria/guarani/hexagonal/guarani/alumnos/alumno/domain/ports/in/CreatePersonalesByNroDocumentoUseCase.java
@@ -1,6 +1,10 @@
 package um.tesoreria.guarani.hexagonal.guarani.alumnos.alumno.domain.ports.in;
 
+import um.tesoreria.guarani.hexagonal.guarani.alumnos.alumno.infrastructure.client.dto.CreatePersonalesResponse;
+
+import java.util.List;
+
 public interface CreatePersonalesByNroDocumentoUseCase {
 
-    Boolean createPersonalesByNroDocumento(String nroDocumento);
+    List<CreatePersonalesResponse> createPersonalesByNroDocumento(String nroDocumento);
 }