+++ src/main/java/um/tesoreria/guarani/hexagonal/guarani/alumnos/alumno/infrastructure/client/adapter/CreatePersonalesAdapter.java
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Component;
 import um.tesoreria.guarani.hexagonal.guarani.alumnos.alumno.domain.model.AlumnoGuarani;
 import um.tesoreria.guarani.hexagonal.guarani.alumnos.alumno.domain.ports.out.CreatePersonalesPort;
 import um.tesoreria.guarani.hexagonal.guarani.alumnos.alumno.infrastructure.client.AlumnoGuaraniClient;
+import um.tesoreria.guarani.hexagonal.guarani.alumnos.alumno.infrastructure.client.dto.CreatePersonalesResponse;
 
 @Component
 @RequiredArgsConstructor
@@ -13,7 +14,7 @@ public class CreatePersonalesAdapter implements CreatePersonalesPort {
     private final AlumnoGuaraniClient alumnoGuaraniClient;
 
     @Override
-    public Boolean createPersonales(AlumnoGuarani alumno) {
+    public CreatePersonalesResponse createPersonales(AlumnoGuarani alumno) {
         return alumnoGuaraniClient.createPersonales(alumno);
     }
 }